feat：升级jar版本

+++ src/main/java/com/tangerineSpecter/oms/system/domain/dto/user/CardNoteInfoDto.java
@@ -22,7 +22,7 @@ public class CardNoteInfoDto {
     @ApiModelProperty("笔记数量")
     private int noteCount;
     @ApiModelProperty("标签数量")
-    private int tagCount;
+    private long tagCount;
     @ApiModelProperty("记录天数")
     private int day;
     @ApiModelProperty("提交记录信息")
